+++ demosplan/DemosPlanCoreBundle/Controller/Faq/FaqApiController.php
@@ -12,7 +12,7 @@ namespace demosplan\DemosPlanCoreBundle\Controller\Faq;
 
 use DemosEurope\DemosplanAddon\Controller\APIController;
 use DemosEurope\DemosplanAddon\Response\APIResponse;
-use demosplan\DemosPlanCoreBundle\Annotation\DplanPermissions;
+use demosplan\DemosPlanCoreBundle\Attribute\DplanPermissions;
 use demosplan\DemosPlanCoreBundle\Logic\Faq\FaqHandler;
 use demosplan\DemosPlanCoreBundle\Logic\Logger\ApiLogger;
 use Exception;
@@ -22,10 +22,9 @@ use Symfony\Component\Routing\Attribute\Route;
 class FaqApiController extends APIController
 {
     /**
-     * @DplanPermissions("area_admin_faq")
-     *
      * @deprecated use `api_resource_update` route instead
      */
+    #[DplanPermissions('area_admin_faq')]
     #[Route(path: '/api/1.0/faq/{faqId}', methods: ['PATCH'], name: 'dp_api_admin_faq_update', options: ['expose' => true])]
     public function update(ApiLogger $apiLogger, string $faqId): Response
     {
@@ -44,9 +43,7 @@ class FaqApiController extends APIController
         }
     }
 
-    /**
-     * @DplanPermissions("area_admin_faq")
-     */
+    #[DplanPermissions('area_admin_faq')]
     #[Route(path: '/api/1.0/faq/{faqId}', methods: ['DELETE'], name: 'dp_api_admin_faq_delete', options: ['expose' => true])]
     public function delete(string $faqId, FaqHandler $faqHandler): APIResponse
     {