+++ demosplan/DemosPlanCoreBundle/Controller/User/DemosPlanUserAuthenticationController.php
@@ -13,7 +13,7 @@ namespace demosplan\DemosPlanCoreBundle\Controller\User;
 use DateTime;
 use DemosEurope\DemosplanAddon\Contracts\CurrentUserInterface;
 use DemosEurope\DemosplanAddon\Contracts\PermissionsInterface;
-use demosplan\DemosPlanCoreBundle\Annotation\DplanPermissions;
+use demosplan\DemosPlanCoreBundle\Attribute\DplanPermissions;
 use demosplan\DemosPlanCoreBundle\Entity\User\AnonymousUser;
 use demosplan\DemosPlanCoreBundle\Entity\User\User;
 use demosplan\DemosPlanCoreBundle\Exception\InvalidArgumentException;
@@ -77,12 +77,11 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
     /**
      * Passwort ändern.
      *
-     * @DplanPermissions("area_mydata_password")
-     *
      * @return Response
      *
      * @throws Exception
      */
+    #[DplanPermissions('area_mydata_password')]
     #[Route(name: 'DemosPlan_user_change_password', path: '/password/change', options: ['expose' => true])]
     public function changePassword(Request $request): RedirectResponse
     {
@@ -104,12 +103,11 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
      * Request change of email.
      * Send Mail to verify change of E-Mail-Address.
      *
-     * @DplanPermissions("feature_change_own_email")
-     *
      * @return RedirectResponse|Response
      *
      * @throws Exception
      */
+    #[DplanPermissions('feature_change_own_email')]
     #[Route(name: 'DemosPlan_user_change_email_request', path: '/email/change')]
     public function changeEmailRequest(Request $request, PasswordHasherFactoryInterface $hasherFactory): RedirectResponse
     {
@@ -127,7 +125,7 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
     }
 
     #[Route(path: '/authentication/2fa/qr-code', name: 'DemosPlan_user_qr_code')]
-    #[\demosplan\DemosPlanCoreBundle\Attribute\DplanPermissions('feature_2fa')]
+    #[DplanPermissions('feature_2fa')]
     public function displayGoogleAuthenticatorQrCode(BuilderInterface $builder, TotpAuthenticatorInterface $totpAuthenticator): QrCodeResponse
     {
         $qrCodeContent = $totpAuthenticator->getQRContent($this->getUser());
@@ -142,7 +140,7 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
     }
 
     #[Route(path: '/authentication/2fa/enable', name: 'DemosPlan_user_2fa_enable')]
-    #[\demosplan\DemosPlanCoreBundle\Attribute\DplanPermissions('feature_2fa')]
+    #[DplanPermissions('feature_2fa')]
     public function enable2fa(
         CurrentUserInterface $currentUser,
         EntityManagerInterface $entityManager,
@@ -158,7 +156,7 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
     }
 
     #[Route(path: '/authentication/2faemail/enable', name: 'DemosPlan_user_2fa_email_enable')]
-    #[\demosplan\DemosPlanCoreBundle\Attribute\DplanPermissions('feature_2fa')]
+    #[DplanPermissions('feature_2fa')]
     public function enable2faemail(
         CodeGeneratorInterface $codeGenerator,
         CurrentUserInterface $currentUser,
@@ -172,7 +170,7 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
     }
 
     #[Route(path: '/authentication/2faemail/send', name: 'DemosPlan_user_2fa_email_send')]
-    #[\demosplan\DemosPlanCoreBundle\Attribute\DplanPermissions('feature_2fa')]
+    #[DplanPermissions('feature_2fa')]
     public function send2faemail(
         CodeGeneratorInterface $codeGenerator,
         CurrentUserInterface $currentUser,
@@ -184,9 +182,8 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
 
     /**
      * Set email address of user. Called via link which was sent to user via email.
-     *
-     * @DplanPermissions("feature_change_own_email")
      */
+    #[DplanPermissions('feature_change_own_email')]
     #[Route(name: 'DemosPlan_user_doubleoptin_change_email', path: 'email/change/doubleoptin/{uId}/{key}')]
     public function changeEmailConfirmation(string $uId, string $key): RedirectResponse
     {
@@ -217,12 +214,11 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
     }
 
     /**
-     *  @DplanPermissions({"area_demosplan","feature_password_recovery"})
-     *
      * @return RedirectResponse|Response
      *
      * @throws MessageBagException
      */
+    #[DplanPermissions(['area_demosplan', 'feature_password_recovery'])]
     #[Route(name: 'DemosPlan_user_password_recover', path: '/password/recover', options: ['expose' => true])]
     public function recoverPassword(RateLimiterFactory $userRegisterLimiter, Request $request)
     {
@@ -295,12 +291,11 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
     /**
      * Alternatives Loginform auf einer ganzen Seite.
      *
-     * @DplanPermissions("area_demosplan")
-     *
      * @return Response
      *
      * @throws AccessDeniedException|Exception
      */
+    #[DplanPermissions('area_demosplan')]
     #[Route(name: 'DemosPlan_user_login_alternative', path: '/dplan/login', options: ['expose' => true])]
     public function alternativeLogin(
         CacheInterface $cache,
@@ -390,12 +385,11 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
     /**
      * Dislay logout landing page.
      *
-     * @DplanPermissions("area_demosplan")
-     *
      * @return RedirectResponse|Response
      *
      * @throws MessageBagException
      */
+    #[DplanPermissions('area_demosplan')]
     #[Route(name: 'DemosPlan_user_logout_success', path: '/user/logout/success')]
     public function logoutSuccess(PermissionsInterface $permissions)
     {
@@ -411,12 +405,11 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
     }
 
     /**
-     * @DplanPermissions("area_demosplan")
-     *
      * @return RedirectResponse|Response
      *
      * @throws Exception
      */
+    #[DplanPermissions('area_demosplan')]
     #[Route(name: 'DemosPlan_user_doubleoptin_invite_confirmation', path: '/doubleoptin/{uId}/{token}')]
     public function confirmInvitation(UserHasher $userHasher, string $token, string $uId)
     {
@@ -437,12 +430,11 @@ class DemosPlanUserAuthenticationController extends DemosPlanUserController
     }
 
     /**
-     * @DplanPermissions("area_demosplan")
-     *
      * @return RedirectResponse|Response
      *
      * @throws Exception
      */
+    #[DplanPermissions('area_demosplan')]
     #[Route(name: 'DemosPlan_user_password_set', path: '/user/{uId}/setpass/{token}', options: ['expose' => true])]
     public function setPassword(
         FlashMessageHandler $flashMessageHandler,